+++ script/Process/__main__.py
@@ -1,4 +1,4 @@
-from process import * 
+from processPerBrand import * 
 
 if __name__ == "__main__" :
     congif = getconfig()
